backend/pyproject.toml
[tool.poetry]
name = "vosu-backend"
version = "0.1.0"
description = "FastAPI + Celery API для VOSU"
authors = ["Роман <roman@example.com>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.10"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
torch = ">=2.7.0,<3.0.0"
openai-whisper = ">=20240930,<20240931"
omegaconf = ">=2.3.0,<3.0.0"
openai = ">=1.82.0,<2.0.0"
redis = {extras = ["async"], version = "^6.2.0"}
celery = ">=5.5.2,<6.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
python-multipart = ">=0.0.20,<0.0.21"
hydra-core = ">=1.3.2,<2.0.0"
lightning = ">=2.5.1.post0,<3.0.0"
lhotse = ">=1.30.3,<2.0.0"
einops = ">=0.8.1,<0.9.0"
transformers = ">=4.52.3,<5.0.0"
sentencepiece = ">=0.2.0,<0.3.0"
pandas = ">=2.2.3,<3.0.0"
jiwer = "^3.1.0"
librosa = "^0.11.0"
pyannote-core = "^5.0.0"
braceexpand = "^0.1.7"
nemo-toolkit = "^2.3.1"
webdataset = "^0.2.111"
datasets = "^3.6.0"
editdistance = "^0.8.1"
pyannote-metrics = "^3.2.1"
ipython = ">=8.0,<9.0"
aiofiles = "^24.1.0"

[tool.poetry.dev-dependencies]
# black = ">=23.1.0"
# isort = ">=5.10.0"
# mypy = ">=1.5.0"

[[tool.poetry.packages]]
include = "app"
from = "."

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
